+++ Movie List/Models/Movie.swift
@@ -7,4 +7,5 @@ import Foundation
 
 struct Movie {
     let movieName: String
+    var seen: Bool = false
 }
